--- res/drawable/ic_security_24dp.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<bitmap xmlns:android="http://schemas.android.com/apk/res/android"
-    android:src="@drawable/ic_security_white_24dp"
-    android:tint="?attr/pref_icon_tint"/>
