--- res/xml/preferences_app_protection.xml
@@ -1,42 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
-    <org.thoughtcrime.securesms.components.SwitchPreferenceCompat
-        android:defaultValue="false"
-        android:key="pref_screen_security"
-        android:summary="@string/pref_screen_security_explain"
-        android:title="@string/pref_screen_security" />
-
-    <org.thoughtcrime.securesms.components.SwitchPreferenceCompat
-        android:defaultValue="false"
-        android:key="pref_incognito_keyboard"
-        android:summary="@string/pref_incognito_keyboard_explain"
-        android:title="@string/pref_incognito_keyboard" />
-
-    <PreferenceCategory android:key="pref_android_screen_lock_category" android:title="@string/screenlock_title">
-
-        <org.thoughtcrime.securesms.components.SwitchPreferenceCompat
-            android:defaultValue="false"
-            android:key="pref_android_screen_lock"
-            android:summary="@string/screenlock_explain"
-            android:title="@string/screenlock_title" />
-
-        <org.thoughtcrime.securesms.components.SwitchPreferenceCompat
-            android:defaultValue="false"
-            android:key="pref_timeout_passphrase"
-            android:summary="@string/screenlock_inactivity_timeout_explain"
-            android:title="@string/screenlock_inactivity_timeout" />
-
-        <Preference
-            android:dependency="pref_timeout_passphrase"
-            android:key="pref_timeout_interval"
-            android:title="@string/screenlock_inactivity_timeout_interval" />
-
-        <Preference
-            android:key="pref_change_passphrase"
-            android:summary="@string/pref_change_secret_explain"
-            android:title="@string/pref_change_secret" />
-
-    </PreferenceCategory>
-
-</PreferenceScreen>
