+++ src/org/thoughtcrime/securesms/BaseActionBarActivity.java
@@ -1,36 +1,25 @@
 package org.thoughtcrime.securesms;
 
-import android.content.Intent;
 import android.os.Build;
 import android.os.Bundle;
-import androidx.annotation.NonNull;
-import androidx.core.app.ActivityCompat;
-import androidx.core.app.ActivityOptionsCompat;
-import androidx.appcompat.app.AppCompatActivity;
 import android.util.Log;
 import android.view.KeyEvent;
 import android.view.Menu;
 import android.view.MenuItem;
-import android.view.View;
 import android.view.ViewConfiguration;
 import android.view.WindowManager;
-import android.widget.Toast;
+
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
 
 import org.thoughtcrime.securesms.util.Prefs;
-import org.thoughtcrime.securesms.util.ScreenLockUtil;
 
 import java.lang.reflect.Field;
-import java.util.Timer;
 
 
 public abstract class BaseActionBarActivity extends AppCompatActivity {
 
   private static final String TAG = BaseActionBarActivity.class.getSimpleName();
-
-  private Timer timer;
-
-  private boolean isWaitingForResult;
-  private boolean isHiddenByScreenLock;
 
   @Override
   protected void onCreate(Bundle savedInstanceState) {
@@ -40,67 +29,10 @@ public abstract class BaseActionBarActivity extends AppCompatActivity {
     super.onCreate(savedInstanceState);
   }
 
-    @Override
-    protected void onStart() {
-        if (ScreenLockUtil.isScreenLockEnabled(this) && ScreenLockUtil.getShouldLockApp() && !isWaitingForResult) {
-          ScreenLockUtil.applyScreenLock(this);
-        } else if (isHiddenByScreenLock) {
-          findViewById(android.R.id.content).setVisibility(View.VISIBLE);
-          isHiddenByScreenLock = false;
-        }
-        super.onStart();
-    }
-
-  @Override
-  protected void onStop() {
-    if (ScreenLockUtil.isScreenLockEnabled(this)) {
-      findViewById(android.R.id.content).setVisibility(View.GONE);
-      isHiddenByScreenLock = true;
-    }
-    super.onStop();
-  }
-
-  @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-      super.onActivityResult(requestCode, resultCode, data);
-      isWaitingForResult = false;
-      if (requestCode == ScreenLockUtil.REQUEST_CODE_CONFIRM_CREDENTIALS) {
-            if (resultCode == RESULT_OK) {
-                ScreenLockUtil.setShouldLockApp(false);
-            } else {
-                Toast.makeText(this, R.string.screenlock_authentication_failed, Toast.LENGTH_SHORT).show();
-                ScreenLockUtil.applyScreenLock(this);
-            }
-        }
-    }
-
   @Override
   protected void onResume() {
     super.onResume();
     initializeScreenshotSecurity();
-    initializeScreenLockTimeout();
-  }
-
-  private void initializeScreenLockTimeout() {
-    if (ScreenLockUtil.isScreenLockTimeoutEnabled(this)) {
-        timer = ScreenLockUtil.scheduleScreenLockTimer(timer, this);
-    }
-  }
-
-  @Override
-    protected void onPause() {
-      super.onPause();
-      tearDownScreenLockTimeout();
-  }
-
-  private void tearDownScreenLockTimeout() {
-    ScreenLockUtil.cancelScreenLockTimer(timer);
-  }
-
-  @Override
-  public void onUserInteraction() {
-    super.onUserInteraction();
-    initializeScreenLockTimeout();
   }
 
   @Override
@@ -142,14 +74,6 @@ public abstract class BaseActionBarActivity extends AppCompatActivity {
       Log.w(TAG, "Failed to force overflow menu.");
     } catch (NoSuchFieldException e) {
       Log.w(TAG, "Failed to force overflow menu.");
-    }
-  }
-
-  @Override
-  public void startActivityForResult(Intent intent, int requestCode) {
-    super.startActivityForResult(intent, requestCode);
-    if (requestCode != -1) {
-      isWaitingForResult = true;
     }
   }
